requirements.txt
numpy
torch
torchvision
torchmetrics[image]
tqdm
wandb
pyyaml